hairgain/app/services/image_processor.py
import os
import uuid
from PIL import Image
import numpy as np
from pathlib import Path

def process_images(face_path, hair_shape_path, hair_color_path, auto_align=True, blend_threshold=0.5):
    """
    Process the uploaded images and generate a customized hair visualization
    
    Args:
        face_path (str): Path to the uploaded face image
        hair_shape_path (str): Path to the hair shape image
        hair_color_path (str): Path to the hair color image
        auto_align (bool): Whether to auto-align the images
        blend_threshold (float): Threshold value for blending (0.0 to 1.0)
        
    Returns:
        str: URL of the processed result image
    """
    try:
        # Load images
        face_img = Image.open(face_path)
        hair_shape_img = Image.open(hair_shape_path)
        hair_color_img = Image.open(hair_color_path)
        
        # Resize images to ensure they're compatible
        target_size = (512, 512)
        face_img = face_img.resize(target_size)
        hair_shape_img = hair_shape_img.resize(target_size)
        hair_color_img = hair_color_img.resize(target_size)
        
        # Convert to numpy arrays for processing
        face_array = np.array(face_img)
        hair_shape_array = np.array(hair_shape_img)
        hair_color_array = np.array(hair_color_img)
        
        # Apply hair shape as a mask and color from the hair color image
        # This is a simplified example - real implementation would be more complex
        # The blend_threshold controls how strongly the hair is blended with the face
        # Higher threshold means more pronounced effect
        # auto_align determines whether facial landmarks should be aligned
        
        # Create result directory if it doesn't exist
        result_dir = Path('static/results')
        result_dir.mkdir(exist_ok=True, parents=True)
        
        # Generate a unique filename for the result
        result_filename = f"{uuid.uuid4()}.png"
        result_path = result_dir / result_filename
        
        # Save the processed image
        # For now, we'll just save the face image as a placeholder
        # In a real implementation, this would be the combined result using blend_threshold and auto_align
        face_img.save(result_path)
        
        # Return the URL to the result image
        return f"/static/results/{result_filename}"
    
    except Exception as e:
        print(f"Error processing images: {str(e)}")
        raise 