hairgain/requirements.txt
addict==2.4.0
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.5.0
aniso8601 @ file:///home/conda/feedstock_root/build_artifacts/aniso8601_1737511056219/work
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.13.3
bezier==2024.6.20
bleach==6.2.0
blinker==1.9.0
botbuilder==0.0.1
build==1.2.2.post1
cached-property==2.0.1
cachetools==5.5.1
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
chromedriver-autoinstaller==0.6.4
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==4.0.0
codeium-jupyter==1.1.21
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.10
contourpy==1.2.0
crc32c==2.4
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.2.7.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
dlib @ file:///D:/bld/dlib-split_1726092764993/work
dnspython==2.7.0
docker-pycreds==0.4.0
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
exchangelib==5.5.1
executing==2.2.0
face_alignment==1.3.4
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.13.1
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
flask-restx @ file:///tmp/build/80754af9/flask-restx_1635786669713/work
Flask-Script==2.0.6
Flask-SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/flask-sqlalchemy_1616102249273/work
Flask-Testing==0.8.1
flatbuffers==25.2.10
fonttools==4.50.0
fpie==0.2.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.2.0
ftfy==6.2.0
gdown==3.12.2
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.70.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
instagrapi==2.1.3
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.2
isodate==0.7.2
isoduration==20.11.0
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
jedi==0.19.2
Jinja2==3.1.3
jiter==0.9.0
joblib==1.2.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications @ file:///tmp/tmpk0f344m9/src
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.13
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==32.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lazy_loader==0.3
llvmlite==0.42.0
lxml==5.3.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.32.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.2.7
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
notebook==7.0.8
notebook_shim==0.2.4
numba==0.59.1
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.61.1
opencv-python==4.7.0.72
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
phonenumbers==8.13.54
Pillow==10.0.0
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
platformdirs==4.3.6
posthog==3.12.0
premailer==3.10.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.6.0
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline==2.1
pyreadline3==3.5.4
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==1.0.1
PySocks==1.7.1
pyspnego==0.11.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-docx==1.1.0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-pptx==0.6.23
python-telegram-bot==21.10
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==26.2.1
qtconsole==5.5.1
QtPy==2.4.1
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
regex==2023.12.25
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.28.1
selenium-stealth==1.0.6
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
slack_sdk==3.35.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.27
sspilib==0.2.0
stack-data==0.6.3
starlette==0.45.3
streamlit==1.43.2
sympy==1.13.1
taichi==1.7.3
Telethon==1.38.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
torch==2.0.0+cu118
torchmetrics==0.11.4
torchvision==0.15.1+cu118
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.36.2
trio==0.28.0
trio-websocket==0.11.1
twilio==9.4.4
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.3.1
ujson==5.10.0
undetected-chromedriver==3.5.5
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.0
waitress==3.0.2
wandb==0.15.10
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1644332431572/work
widgetsnbextension==4.0.10
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.2
yagmail==0.15.293
yarl==1.9.4
zipp==3.17.0
